requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
Pillow==10.1.0
qrcode[pil]==7.4.2
gunicorn==21.2.0
whitenoise==6.6.0
